refactor endpoint ssrf tests
